requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2022.9.24
channels==4.0.0
charset-normalizer==2.1.1
Django==4.1.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-rest-authemail==2.1.4
django-summernote==0.8.20.0
djangorestframework==3.14.0
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.2.0
PyMySQL==1.0.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
webencodings==0.5.1
